Simplify notebooks: no workaround needed with lcg85_swan3
The pytimber notebooks contained workarounds due to the fact that
the package was not fully integrated in the lcg releases.
Starting from lcg85_swan3 this is not the case anymore and the
narrative can be made simpler and more expressive.
SWAN is ready for analytics of data coming from the LHC machine.